Update frontpage
Use list instead of search as the frontpage
